requirements.txt
asgiref==3.4.1
autopep8==1.5.7
dj-database-url==0.5.0
Django==3.2.5
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2-binary==2.9.3
pycodestyle==2.7.0
python-decouple==3.6
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
whitenoise==6.0.0
